Split snapshot and annotation tests into separate file
Move snapshot tests and annotation tests from tests.rs to
snapshot_tests.rs to keep tests.rs under the 1000-line limit.
Clean up stale snapshot files from the old module path.
